+++ lib/src/editor/find_replace_menu/find_replace_widget.dart
@@ -161,7 +161,7 @@ class _FindMenuState extends State<FindMenu> {
 
   final findTextEditingController = TextEditingController();
 
-  String queriedPattern = '';
+  String message = AppFlowyEditorLocalizations.current.emptySearchBoxHint;
 
   bool showReplaceMenu = false;
   bool caseSensitive = false;
@@ -250,9 +250,7 @@ class _FindMenuState extends State<FindMenu> {
           padding: const EdgeInsets.symmetric(horizontal: 8.0),
           alignment: Alignment.centerLeft,
           child: Text(
-            matches.isEmpty
-                ? widget.localizations?.noResult ?? 'No Result'
-                : '$selectedIndex of ${matches.length}',
+            matches.isEmpty ? message : '$selectedIndex of ${matches.length}',
           ),
         ),
         // previous match button
@@ -322,11 +320,25 @@ class _FindMenuState extends State<FindMenu> {
   }
 
   void _searchPattern() {
-    if (findTextEditingController.text.isEmpty) {
-      return;
+    String error;
+
+    // the following line needs to be executed even if
+    // findTextEditingController.text.isEmpty, otherwise the previous
+    // matches will persist
+    error =
+        widget.searchService.findAndHighlight(findTextEditingController.text);
+
+    switch (error) {
+      case 'Regex':
+        message = AppFlowyEditorLocalizations.current.regexError;
+      case 'Empty':
+        message = AppFlowyEditorLocalizations.current.emptySearchBoxHint;
+      default:
+        message = widget.localizations?.noResult ??
+            AppFlowyEditorLocalizations.current.noFindResult;
     }
-    widget.searchService.findAndHighlight(findTextEditingController.text);
-    setState(() => queriedPattern = findTextEditingController.text);
+
+    _setState();
   }
 
   void _setState() {