+++ lib/src/editor/toolbar/mobile/mobile_floating_toolbar/mobile_floating_toolbar.dart
@@ -25,7 +25,11 @@ class MobileFloatingToolbar extends StatefulWidget {
   final EditorState editorState;
   final EditorScrollController editorScrollController;
   final Widget child;
-  final Widget Function(BuildContext context, Offset anchor) toolbarBuilder;
+  final Widget Function(
+    BuildContext context,
+    Offset anchor,
+    Function closeToolbar,
+  ) toolbarBuilder;
 
   @override
   State<MobileFloatingToolbar> createState() => _MobileFloatingToolbarState();
@@ -94,12 +98,14 @@ class _MobileFloatingToolbarState extends State<MobileFloatingToolbar>
     } else if (selection.isCollapsed) {
       if (_isToolbarVisible) {
         _clear();
-      } else if (prevSelection == selection) {
+      } else if (prevSelection == selection &&
+          editorState.selectionUpdateReason == SelectionUpdateReason.uiEvent) {
         _showAfterDelay(const Duration(milliseconds: 400));
       }
       prevSelection = selection;
     } else {
       // uses debounce to avoid the computing the rects too frequently.
+      _clear();
       _showAfterDelay(const Duration(milliseconds: 400));
     }
   }
@@ -153,7 +159,11 @@ class _MobileFloatingToolbarState extends State<MobileFloatingToolbar>
     BuildContext context,
     Offset offset,
   ) {
-    return widget.toolbarBuilder(context, offset);
+    return widget.toolbarBuilder(
+      context,
+      offset,
+      _clear,
+    );
   }
 
   Rect _findSuitableRect(Iterable<Rect> rects) {