+++ lib/src/editor/toolbar/mobile/toolbar_items/color/text_color_options_widgets.dart
@@ -1,13 +1,13 @@
-import 'package:flutter/material.dart';
 import 'package:appflowy_editor/appflowy_editor.dart';
+import 'package:flutter/material.dart';
 
 class TextColorOptionsWidgets extends StatefulWidget {
   const TextColorOptionsWidgets(
     this.editorState,
     this.selection, {
     this.textColorOptions,
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   final Selection selection;
   final EditorState editorState;
@@ -21,7 +21,7 @@ class TextColorOptionsWidgets extends StatefulWidget {
 class _TextColorOptionsWidgetsState extends State<TextColorOptionsWidgets> {
   @override
   Widget build(BuildContext context) {
-    final style = MobileToolbarStyle.of(context);
+    final style = MobileToolbarTheme.of(context);
 
     final selection = widget.selection;
     final nodes = widget.editorState.getNodesInSelection(selection);
@@ -80,7 +80,7 @@ class _TextColorOptionsWidgetsState extends State<TextColorOptionsWidgets> {
               },
               isSelected: isSelected,
             );
-          }).toList(),
+          }),
         ],
       ),
     );