+++ lib/src/flutter/overlay.dart
@@ -243,10 +243,10 @@ class Overlay extends StatefulWidget {
   /// Rather than creating an overlay, consider using the overlay that is
   /// created by the [Navigator] in a [WidgetsApp] or a [MaterialApp] for the application.
   const Overlay({
-    Key? key,
+    super.key,
     this.initialEntries = const <OverlayEntry>[],
     this.clipBehavior = Clip.hardEdge,
-  }) : super(key: key);
+  });
 
   /// The entries to include in the overlay initially.
   ///
@@ -574,13 +574,11 @@ class OverlayState extends State<Overlay> with TickerProviderStateMixin {
 /// The first [skipCount] children are considered "offstage".
 class _Theatre extends MultiChildRenderObjectWidget {
   const _Theatre({
-    Key? key,
     this.skipCount = 0,
     this.clipBehavior = Clip.hardEdge,
-    List<Widget> children = const <Widget>[],
+    super.children,
   })  : assert(skipCount >= 0),
-        assert(children.length >= skipCount),
-        super(key: key, children: children);
+        assert(children.length >= skipCount);
 
   final int skipCount;
 
@@ -614,7 +612,7 @@ class _Theatre extends MultiChildRenderObjectWidget {
 }
 
 class _TheatreElement extends MultiChildRenderObjectElement {
-  _TheatreElement(_Theatre widget) : super(widget);
+  _TheatreElement(_Theatre super.widget);
 
   @override
   _RenderTheatre get renderObject => super.renderObject as _RenderTheatre;