+++ CHANGELOG.md
@@ -1,3 +1,34 @@
+## 2.0.0
+* feat: support Flutter stable channel by @LucasXu0 in ([#581](https://github.com/AppFlowy-IO/appflowy-editor/pull/581))
+* feat: support toggling format style when the selection is collapsed by @LucasXu0 in ([#586](https://github.com/AppFlowy-IO/appflowy-editor/pull/586))
+* feat: optimize editing experience on mobile by @LucasXu0 in ([#592](https://github.com/AppFlowy-IO/appflowy-editor/pull/592))
+* feat: support display number of the numbered list in latin or roman format by @q200892907 in ([#595](https://github.com/AppFlowy-IO/appflowy-editor/pull/595))
+* fix: don't rebuild mobile toolbar every time by @LucasXu0 in ([#579](https://github.com/AppFlowy-IO/appflowy-editor/pull/579))
+* fix: solve the left and right key bug by @q200892907 in ([#584](https://github.com/AppFlowy-IO/appflowy-editor/pull/584))
+* fix: unable to set toolbar icon theme by @LucasXu0 in ([#596](https://github.com/AppFlowy-IO/appflowy-editor/pull/596))
+
+
+## 2.0.0-beta.1
+* feat: support Flutter stable channel by @LucasXu0 in ([#581](https://github.com/AppFlowy-IO/appflowy-editor/pull/581))
+
+## 1.5.2
+* feat: show placeholder override for paragraph block by @Xazin in ([#571](https://github.com/AppFlowy-IO/appflowy-editor/pull/571))
+* feat: Checkbox unresponsive in custom Theme example in editor example app by @johansutrisno in ([#543](https://github.com/AppFlowy-IO/appflowy-editor/pul/543))
+* fix: quit editing button doesn't work by @LucasXu0 in ([#5677](https://github.com/AppFlowy-IO/appflowy-editor/pull/577))
+* fix: check if header exists for footer by @MayurSMahajan in ([#559](https://github.com/AppFlowy-IO/appflowy-editor/pull/559))
+* fix: 'PageStorage.of' returns null by @LucasXu0 in ([#575](https://github.com/AppFlowy-IO/appflowy-editor/pull/575))
+* fix: catch regex `FormatException` and show a corresponding error message  by @sun-jiao in ([#573](https://github.com/AppFlowy-IO/appflowy-editor/pull/573))
+* chore: changed padding values and added borderRadius by @Kritarthsharma in ([#567](https://github.com/AppFlowy-IO/appflowy-editor/pull/567))
+* chore: refactor character replacement by @Xazin in ([#570](https://github.com/AppFlowy-IO/appflowy-editor/pull/570))
+
+## 1.5.1
+* feat: toggledStyle notifier by @Xazin in ([#562](https://github.com/AppFlowy-IO/appflowy-editor/pull/562))
+* feat: power checkbox shortcut by @Xazin in ([#553](https://github.com/AppFlowy-IO/appflowy-editor/pull/553))
+* * fix: Applied Corner Radius to Block Selection Area by @0xJaskeerat in ([#557](https://github.com/AppFlowy-IO/appflowy-editor/pull/557))
+* fix: do not select all if selection is null by @Xazin in ([#563](https://github.com/AppFlowy-IO/appflowy-editor/pull/563))
+* fix: on ui event reset selection type by @Xazin in ([#55](https://github.com/AppFlowy-IO/appflowy-editor/pull/555))
+* fix: image copy error width/height is double by @q200892907 in ([#566](https://github.com/AppFlowy-IO/appflowy-editor/pull/566))
+
 ## 1.5.0
 * feat: enable toggling and canceling of formatting with shortcuts by @LucasXu0 in ([#528](https://github.com/AppFlowy-IO/appflowy-editor/pull/528))
 * feat: disable the toolbar in the code block by @LucasXu0 in ([#534](https://github.com/AppFlowy-IO/appflowy-editor/pull/534))